Docs: add autogenerated rules table to README (#110)
This change redesigns the rules and configurations sections in the README.

The rules section is now an autogenerated table with the complete list of rules and emojis to indicate:
* What configuration (recommended or two) each rule belongs to
* Whether each rule is autofixable

A new script `yarn update` will automatically regenerate the rules table. I have also added a CI test to ensure the rules table is always correct and up-to-date (replacing the previous README tests).

The configurations section has also been converted into a table directly above the rules table for easy reference.

This README design and script is based on the same ones used by [eslint-plugin-ember](https://github.com/ember-cli/eslint-plugin-ember) and [ember-template-lint](https://github.com/ember-template-lint/ember-template-lint).